Use tqdm() library and remove pytorch_classification Bar
