+++ Arena.py
@@ -1,7 +1,6 @@
 import logging
-import time
 
-from pytorch_classification.utils import Bar, AverageMeter
+from tqdm import tqdm
 
 log = logging.getLogger(__name__)
 
@@ -73,17 +72,12 @@ class Arena():
             twoWon: games won by player2
             draws:  games won by nobody
         """
-        eps_time = AverageMeter()
-        bar = Bar('Arena.playGames', max=num)
-        end = time.time()
-        eps = 0
-        maxeps = int(num)
 
         num = int(num / 2)
         oneWon = 0
         twoWon = 0
         draws = 0
-        for _ in range(num):
+        for _ in tqdm(range(num), desc="Arena.playGames (1)"):
             gameResult = self.playGame(verbose=verbose)
             if gameResult == 1:
                 oneWon += 1
@@ -91,20 +85,10 @@ class Arena():
                 twoWon += 1
             else:
                 draws += 1
-            # bookkeeping + plot progress
-            eps += 1
-            eps_time.update(time.time() - end)
-            end = time.time()
-            bar.suffix = '({eps}/{maxeps}) Eps Time: {et:.3f}s | Total: {total:} | ETA: {eta:}'.format(eps=eps,
-                                                                                                       maxeps=maxeps,
-                                                                                                       et=eps_time.avg,
-                                                                                                       total=bar.elapsed_td,
-                                                                                                       eta=bar.eta_td)
-            bar.next()
 
         self.player1, self.player2 = self.player2, self.player1
 
-        for _ in range(num):
+        for _ in tqdm(range(num), desc="Arena.playGames (2)"):
             gameResult = self.playGame(verbose=verbose)
             if gameResult == -1:
                 oneWon += 1
@@ -112,17 +96,5 @@ class Arena():
                 twoWon += 1
             else:
                 draws += 1
-            # bookkeeping + plot progress
-            eps += 1
-            eps_time.update(time.time() - end)
-            end = time.time()
-            bar.suffix = '({eps}/{maxeps}) Eps Time: {et:.3f}s | Total: {total:} | ETA: {eta:}'.format(eps=eps,
-                                                                                                       maxeps=maxeps,
-                                                                                                       et=eps_time.avg,
-                                                                                                       total=bar.elapsed_td,
-                                                                                                       eta=bar.eta_td)
-            bar.next()
-
-        bar.finish()
 
         return oneWon, twoWon, draws
