--- pytorch_classification/models/imagenet/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import absolute_import
-
-from .resnext import *
