--- pytorch_classification/utils/__init__.py
@@ -1,10 +1,0 @@
-"""Useful utils
-"""
-from .misc import *
-from .logger import *
-from .eval import *
-
-# progress bar
-import os, sys
-sys.path.append(os.path.join(os.path.dirname(__file__), "progress"))
-from progress.bar import Bar as Bar
