--- pytorch_classification/utils/progress/MANIFEST.in
@@ -1,1 +1,0 @@
-include README.rst LICENSE
